Fix calculation of current week and current month for work cycle
Week:
The current time was used in the end date of the week.
This caused the day the be counted as another full workday.
A week now correctly goes from Monday, midnight to next Monday,
midnight.

Month:
The end of the month is now set to the first day of the next month,
midnight, matching how the number of workdays is calculated for
weeks and years (1st of X to 1st of X+1).