WorkCycleData: Initialise weekEnd with weekStart to always get correct month
If weekEnd is initialised with the current date, the calculation breaks when a week
spans two months.
The new month will then be combined with the old months day, resulting in an additional
month added to weekEnd.
E.g. In a week starting on Monday, 30.07 weekEnd was calculated as (30.08 + 7days =  06.09)
from Wednesday on